% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConsensusTME.R
\name{buildConsensusGenes}
\alias{buildConsensusGenes}
\title{Build Consensus Gene Supersets
\code{buildConsensusGenes} Combines pre-processed signatures and filters to create curated gene sets.}
\usage{
buildConsensusGenes(matchedSigs, cancerAll, immuneFilter = NULL,
  stromalFilter = NULL, immuneScore = TRUE)
}
\arguments{
\item{matchedSigs}{list of pre-processed signatures with consistant cell type nomenclature}

\item{cancerAll}{TCGA cancer types to produce gene sets for}

\item{immuneFilter}{list of genes to filter immune gene sets by. If \code{NULL} - \code{methodSignatures$ImmuneGenes}
is used (list of immune genes with negative correlation with tumour purity for each tcga cancer type).
\bold{N.B.} Must have two columns: "Gene_Symbol" & "Cancer".}

\item{stromalFilter}{list of genes to filter stromal gene sets by. If \code{NULL} - \code{methodSignatures$ESTIMATE}
is used (list of ESTIMATE's stromal genes with negative correlation with tumour purity for each tcga cancer type).
\bold{N.B.} Must have two columns: "Gene_Symbol" & "Cancer".}

\item{immuneScore}{logical, when \code{TRUE} (default) an Immune Score is produced representing overall
level of immune infiltration for each sample.}
}
\value{
returns consensusGeneSets
}
\description{
Build Consensus Gene Supersets
\code{buildConsensusGenes} Combines pre-processed signatures and filters to create curated gene sets.
}
