% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConsensusTME.R
\docType{data}
\name{cancerAll}
\alias{cancerAll}
\title{Consensus TME Cancer Types}
\format{An object of class \code{character} of length 32.}
\usage{
data(cancerAll)
}
\description{
character vector containing TCGA abbreviations
}
\details{
TCGA cancer types for which ConsensusTME can generate gene sets for:

ACC BLCA BRCA CESC CHOL COAD DLBC ESCA GBM HNSC KICH KIRC KIRP LGG LIHC LUAD LUSC MESO OV PAAD PCPG PRAD READ SARC SKCM STAD TGCT THCA THYM UCEC UCS UVM.
}
\keyword{datasets}
