% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConsensusTME.R
\docType{data}
\name{consensusGeneSets}
\alias{consensusGeneSets}
\title{Consensus TME Gene Sets}
\format{An object of class \code{list} of length 33.}
\usage{
data(consensusGeneSets)
}
\description{
list of genes for each cell type for each cancer type
}
\details{
Preprocessed ConsensusTME gene sets for each of the TCGA cancer types
}
\keyword{datasets}
