% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConsensusTME.R
\docType{package}
\name{ConsensusTME}
\alias{ConsensusTME}
\alias{ConsensusTME-package}
\title{ConsensusTME: A package for estimation of cell type abundance.}
\description{
ConsensusTME: A package for estimation of cell type abundance.
}
