% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConsensusTME.R
\name{consensusTMEAnalysis}
\alias{consensusTMEAnalysis}
\title{Run ConsensusTME Cell Type Estimation
\code{consensusTMEAnalysis} takes bulk tumour gene expression data and returns
cell type specific enrichment scores for each sample}
\usage{
consensusTMEAnalysis(bulkExp, cancerType, statMethod = c("ssgsea",
  "gsva", "plage", "zscore", "singScore"), singScoreDisp = FALSE,
  immuneScore = TRUE, excludeCells = NULL, parallel.sz = 0,
  parallel.type = "SOCK")
}
\arguments{
\item{bulkExp}{bulk tumour gene expression matrix. HUGO gene symbols as row names & sample IDs
as column names.}

\item{cancerType}{string passed to indicate which cancer type samples are to allow appropriate gene
set to be used. \bold{N.B} samples of different cancer types should be run seperately.}

\item{statMethod}{statistical framework to be used in generating gene set enrichment scores.
These mirror the parameter options of \code{GSVA::gsva()} with the exception of \code{singScore}.
which leverages \code{singscore::multiScore()}. Default: \code{ssgsea}}

\item{singScoreDisp}{Logical, when using singscore method should the dispersion for each
gene set be returned with the enrichment scores. Default: \code{FALSE}}

\item{immuneScore}{logical, when \code{TRUE} (default) an Immune Score is produced representing overall
level of immune infiltration for each sample.}

\item{excludeCells}{a character vector that defines cell types to not generate enrichment scores for.}

\item{parallel.sz}{parameter passed to \code{GSVA::gsva()} - Number of processors to use when doing the calculations in parallel.
This requires to previously load either the parallel or the snow library. If parallel is loaded and this argument is left with its default value (parallel.sz = 0)
then it will use all available core processors unless this is set with a smaller number.
If snow is loaded then this must be set to a positive integer number that specifies the number of processors to employ in the parallel calculation.}

\item{parallel.type}{parameter passed to \code{GSVA::gsva()} - Type of cluster architecture when using snow. \code{"SOCK"} - works on any system (including Windows).
\code{"FORK"} -  faster but should be only used for POSIX systems (Mac, Linux, Unix, BSD).}
}
\value{
returns estimation of cell type abundance for each sample in the bulk tumour gene expression matrix
}
\description{
Run ConsensusTME Cell Type Estimation
\code{consensusTMEAnalysis} takes bulk tumour gene expression data and returns
cell type specific enrichment scores for each sample
}
