% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConsensusTME.R
\name{geneSetEnrichment}
\alias{geneSetEnrichment}
\title{Carry out Gene Set Enrichment Analysis}
\usage{
geneSetEnrichment(bulkExp, signatures, statMethod = c("ssgsea", "gsva",
  "plage", "zscore", "singScore"), singScoreDisp = FALSE,
  parallel.sz = 0, parallel.type = "SOCK")
}
\arguments{
\item{bulkExp}{bulk tumour gene expression matrix. HUGO gene symbols as row names & sample IDs
as column names.}

\item{signatures}{a list with each element containing genes to represent a cell type. The cell types
should be the names of each element of the list.}

\item{statMethod}{statistical framework to be used in generating gene set enrichment scores.
These mirror the parameter options of \code{GSVA::gsva()} with the exception of \code{singScore}.
which leverages \code{singscore::multiScore()}. Default: \code{ssgsea}}

\item{singScoreDisp}{Logical, when using singscore method should the dispersion for each
gene set be returned with the enrichment scores. Default: \code{FALSE}}

\item{parallel.sz}{parameter for \code{GSVA::gsva()} - Number of processors to use when doing the calculations in parallel.
This requires to previously load either the parallel or the snow library. If parallel is loaded and this argument is left with its default value (parallel.sz = 0)
then it will use all available core processors unless this is set with a smaller number.
If snow is loaded then this must be set to a positive integer number that specifies the number of processors to employ in the parallel calculation.}

\item{parallel.type}{parameter for \code{GSVA::gsva()} - Type of cluster architecture when using snow. \code{"SOCK"} - works on any system (including Windows).
\code{"FORK"} -  faster but should be only used for POSIX systems (Mac, Linux, Unix, BSD).}
}
\value{
returns a list with curated signatures ready to be combined
}
\description{
\code{geneSetEnrichment} Combines exisiting signatures and allows additional of a new signature.
}
