% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConsensusTME.R
\name{matchGeneSigs}
\alias{matchGeneSigs}
\title{Combine Existing Signatures Ensuring Consistant Nomenclature}
\usage{
matchGeneSigs(methodSignatures)
}
\arguments{
\item{methodSignatures}{raw gene signatures list generated from methodSignatures.R}
}
\value{
returns a list with curated signatures ready to be combined
}
\description{
\code{matchGeneSigs} Combines existing signatures and allows additional of a new signature.
}
