% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConsensusTME.R
\docType{data}
\name{methodSignatures}
\alias{methodSignatures}
\title{Raw Method Signatures and Filter Data}
\format{list:
\describe{
  \item{Bindea}{Gene sets for 24 immune cells & 2 biological processes}
  \item{Danaher}{Gene sets for 14 immune cells}
  \item{Davoli}{Gene sets for 10 immune cells}
  \item{MCP.Counter}{Gene sets for 10 immune cells & 2 stromal cells}
  \item{xCell}{Gene sets for 34 immune cells, 9 other cells of haematopoietic lineage &
  21 cells of non haematopoietic lineage}
  \item{CIBERSORT}{LM22 signature matrix for 22 immune cell types}
  \item{ImmuneGenes}{Dataframe with immune genes with negative correlation with tumour
  purity for each cancer type}
  \item{ESTIMATE}{Dataframe with ESTIMATES stromal genes that have a negative correlation
  with tumour purity for each cancer type}
}}
\usage{
data(methodSignatures)
}
\description{
Compilation of gene sets, signature matracies and filters for generation
of ConsensusTME gene sets
}
\keyword{datasets}
